feature: Course(과정)을 저장 및 조회할 수 있다.

+++ src/main/java/nextstep/courses/domain/course/CourseRepository.java
@@ -6,5 +6,5 @@ public interface CourseRepository {
 
   Optional<Course> findById(Long id);
 
-  int save(Course course);
+  Long save(Course course);
 }
